+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.56 ] - 2024/03/04
+
+  * Added method `SoundTouchClient.UpdateNowPlayingStatusForSource` to update source-specific NowPlayingStatus object for a given source and sourceAccount value.
+  * Updated model `NowPlayingStatus` to allow creating an instance with specified values.
+  * Replaced `requests` import with `urllib3.request` import in `SoundTouchFirmware` class.
+
 ###### [ 1.0.55 ] - 2023/02/17
 
   * Updated `SoundTouchWebSocket`.`NotifyListeners` method to suppress logging of exception details to the system logger for websocket error events.